requirements.txt
pyserial
#matplotlib
numpy
RPi.GPIO
